requirements.txt
Django==4.2.8
psycopg==3.1.16
psycopg-binary==3.1.16
gunicorn==21.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.2.0
aiogram
requests